+++ test/unit/api_base/test_legacy_aggregation_body.py
@@ -4,18 +4,18 @@
 
 import pytest
 
-from tol.api_base.misc import AggregationBody
-from tol.api_client.exception import BadPostJsonError
+from tol.api_base.misc import LegacyAggregationBody
+from tol.api_client.exception import PostJsonKeyMissingError
 
 
 class TestAggregationBody:
     def test_no_aggregations(self):
         """No aggregations specified throws error"""
-        with pytest.raises(BadPostJsonError):
-            AggregationBody({'irrelevent': 'so?'}).aggs
+        with pytest.raises(PostJsonKeyMissingError):
+            LegacyAggregationBody({'irrelevent': 'so?'}).aggs
 
     def test_good_aggregations(self):
         """Just aggregations, confirm that an integer is returned"""
         agg = {'something': 'here'}
-        parsed = AggregationBody({'aggs': agg})
+        parsed = LegacyAggregationBody({'aggs': agg})
         assert parsed.aggs == {'something': 'here'}