+++ test/unit/api_base/test_operator_config.py
@@ -7,6 +7,7 @@ from tol.core.operator import (
     Aggregator,
     Deleter,
     DetailGetter,
+    LegacyAggregator,
     PageGetter,
     Updater,
     Upserter
@@ -15,9 +16,10 @@ from tol.core.operator.operator_config import DefaultOperatorConfig
 
 
 class _MockDataSource1(
-    DataSource,
     Aggregator,
+    DataSource,
     DetailGetter,
+    LegacyAggregator,
     PageGetter
 ):
     @property
@@ -29,6 +31,9 @@ class _MockDataSource1(
         raise NotImplementedError()
 
     def get_aggregations(*args, **kwargs):
+        raise NotImplementedError()
+
+    def get_aggregations_legacy(*args, **kwargs):
         raise NotImplementedError()
 
     def get_by_id(*args, **kwargs):